// Frontend/src/app/api/model/product.ts
/**
 * ProductCatalog
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Product { 
    id?: number;
    code: string;
    name: string;
    description?: string | null;
    price?: number;
    modified?: string | null;
}